fix: adiciona api error quando a transação não existe na hora de deletar

+++ src/application/use-case/delete-transaction.ts
@@ -1,3 +1,5 @@
+import { ApiError } from '../../common/api-error'
+import { StatusCode } from '../../common/status-code'
 import { DeleteTransactionDTO } from '../../infrastructure/dtos/delete-transaction-dto'
 import { ITransactionRepository } from '../repositories/itransaction-repository'
 
@@ -5,6 +7,12 @@ export class DeleteTransaction {
   constructor(private readonly transactionRepository: ITransactionRepository) {}
 
   async execute(input: DeleteTransactionDTO) {
+    const transactionAlreadyExistis = await this.transactionRepository.get(
+      input.id,
+    )
+    if (!transactionAlreadyExistis) {
+      throw new ApiError('transação não encontrada', StatusCode.NOT_FOUND)
+    }
     return this.transactionRepository.delete(input.id)
   }
 }
